requirements.txt
africastalking==1.2.3
arabic-reshaper==2.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
contextlib2==0.6.0.post1
dj-database-url==0.5.0
Django==4.0.6
django-bootstrap-forms==0.1
django-cors-headers==3.13.0
django-crispy-forms==1.11.0
django-filter==2.4.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jsonpickle==2.2.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.59
requests==2.25.1
schema==0.7.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
